+++ src/main/java/com/gba/pollvote/dto/custom/SessionCustomDTO.java
@@ -1,0 +1,13 @@
+package com.gba.pollvote.dto.custom;
+
+import lombok.*;
+
+@Setter
+@Getter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class SessionCustomDTO {
+    private Long pollId;
+    private Integer sessionDuration;
+}