+++ src/test/java/com/gba/pollvote/service/SessionServiceTest.java
@@ -1,8 +1,13 @@
 package com.gba.pollvote.service;
 
+import com.gba.pollvote.domain.Associate;
 import com.gba.pollvote.domain.Poll;
 import com.gba.pollvote.domain.Session;
+import com.gba.pollvote.domain.Vote;
+import com.gba.pollvote.dto.VoteResultDTO;
 import com.gba.pollvote.repository.PollRepository;
+import com.gba.pollvote.repository.SessionRepository;
+import com.gba.pollvote.repository.VoteRepository;
 import com.gba.pollvote.service.impl.SessionServiceImpl;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -14,6 +19,7 @@ import org.mockito.Mockito;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
+import javax.persistence.EntityNotFoundException;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -24,11 +30,18 @@ import static org.assertj.core.api.Assertions.assertThat;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @AutoConfigureMockMvc
 public class SessionServiceTest {
-    @Mock
-    protected PollRepository pollRepository;
 
     @InjectMocks
     protected SessionServiceImpl sessionService;
+
+    @Mock
+    protected SessionRepository sessionRepository;
+
+    @Mock
+    protected PollRepository pollRepository;
+
+    @Mock
+    protected VoteRepository voteRepository;
 
     @BeforeEach
     public void setUp(){}
@@ -43,8 +56,7 @@ public class SessionServiceTest {
         Session session = Session.builder()
                 .sessionDuration(1)
                 .poll(pollSession.get())
-                .startDate(LocalDateTime.now())
-                .endDate(LocalDateTime.now()).build();
+                .startDate(LocalDateTime.now()).build();
 
         Mockito.when(sessionService.create(session)).thenReturn(session);
         Session expectedAssociate = sessionService.create(session);
@@ -68,5 +80,46 @@ public class SessionServiceTest {
         List<Session> listExpected = sessionService.getAll();
 
         Assertions.assertEquals(listExpected, list);
+    }
+
+    @Test
+    void getVoteResult() {
+        Long id = 1L;
+
+        Session session = Session.builder()
+                .sessionDuration(1)
+                .poll(Poll.builder().name("teste").build())
+                .startDate(LocalDateTime.now()).build();
+        Mockito.when(sessionRepository.findById(id)).thenReturn(Optional.of(session));
+
+        List<Vote> voteList = new ArrayList<>();
+        voteList.add(Vote.builder()
+                        .status(true)
+                        .session(session)
+                        .associate(Associate.builder()
+                                .id(1L).name("Gabriel")
+                                .cpf("04182914201").build()).build());
+
+        Mockito.when(voteRepository.findBySessionId(id)).thenReturn(voteList);
+
+        VoteResultDTO voteResult = sessionService.getSessionResultById(id);
+
+        assertThat(voteResult).isNotNull();
+    }
+    @Test
+    void getVoteResultWithWrongSessionIdException() throws EntityNotFoundException{
+        Long id = 1L;
+
+        Session session = Session.builder()
+                .sessionDuration(1)
+                .poll(Poll.builder().name("teste").build())
+                .startDate(LocalDateTime.now()).build();
+        Mockito.when(sessionRepository.findById(id)).thenReturn(Optional.of(session));
+
+        Assertions.assertThrows(
+                EntityNotFoundException.class,
+                () -> sessionService.getSessionResultById(id)
+        );
+
     }
 }